+++ source/Messaging.Tests/CimMessageAdapter/Messages/RequestChangeOfSupplier/MessageParserTests.cs
@@ -12,12 +12,14 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using System.Xml.Linq;
+using Messaging.CimMessageAdapter.Errors;
 using Messaging.CimMessageAdapter.Messages;
 using Messaging.CimMessageAdapter.Messages.RequestChangeOfSupplier;
 using Messaging.Domain.OutgoingMessages;
@@ -48,6 +50,15 @@ public class MessageParserTests
         };
     }
 
+    public static IEnumerable<object[]> CreateMessagesWithInvalidStructure()
+    {
+        return new List<object[]>
+        {
+            new object[] { CimFormat.Json, CreateInvalidJsonMessage() },
+            new object[] { CimFormat.Xml, CreateInvalidXmlMessage() },
+        };
+    }
+
     [Theory]
     [MemberData(nameof(CreateMessages))]
     public async Task Can_parse(CimFormat format, Stream message)
@@ -73,6 +84,24 @@ public class MessageParserTests
         Assert.Equal("2022-09-07T22:00:00Z", marketActivityRecord.EffectiveDate);
     }
 
+    [Theory]
+    [MemberData(nameof(CreateMessagesWithInvalidStructure))]
+    public async Task Return_error_when_structure_is_invalid(CimFormat format, Stream message)
+    {
+        var result = await _messageParser.ParseAsync(message, format).ConfigureAwait(false);
+
+        Assert.False(result.Success);
+        Assert.Contains(result.Errors, error => error is InvalidMessageStructure);
+    }
+
+    [Fact]
+    public async Task Throw_if_message_format_is_not_known()
+    {
+        var parser = new MessageParser(new List<IMessageParser>());
+
+        await Assert.ThrowsAsync<InvalidOperationException>(() => parser.ParseAsync(CreateXmlMessage(), CimFormat.Xml)).ConfigureAwait(false);
+    }
+
     private static Stream CreateXmlMessage()
     {
         var xmlDoc = XDocument.Load($"cimmessageadapter{Path.DirectorySeparatorChar}messages{Path.DirectorySeparatorChar}xml{Path.DirectorySeparatorChar}RequestChangeOfSupplier.xml");
@@ -86,6 +115,23 @@ public class MessageParserTests
     {
         return ReadTextFile(
             $"cimmessageadapter{Path.DirectorySeparatorChar}messages{Path.DirectorySeparatorChar}json{Path.DirectorySeparatorChar}Request Change of Supplier.json");
+    }
+
+    private static MemoryStream CreateInvalidJsonMessage()
+    {
+        return ReadTextFile($"cimmessageadapter{Path.DirectorySeparatorChar}messages{Path.DirectorySeparatorChar}json{Path.DirectorySeparatorChar}Invalid Request Change of Supplier.json");
+    }
+
+    private static Stream CreateInvalidXmlMessage()
+    {
+        var messageStream = new MemoryStream();
+        using var writer = new StreamWriter(messageStream);
+        writer.Write("This is not XML");
+        writer.Flush();
+        messageStream.Position = 0;
+        var returnStream = new MemoryStream();
+        messageStream.CopyTo(returnStream);
+        return returnStream;
     }
 
     private static MemoryStream ReadTextFile(string path)
